refactor: separate pure functions
