+++ api/src/misc/utils/query-advice.ts
@@ -1,5 +1,5 @@
 import { type Request } from 'express'
-import { hasManyQSearchFields } from '../../datasets/es/commons.js'
+import { hasManyQSearchFields } from '../../datasets/es/operations.ts'
 
 // Builds a short, localized, advisory sentence appended to overload errors (429 compute-budget,
 // 504 "request too long", 429 ES circuit_breaking_exception). It only ever *advises* — it never
